feat(bedrock): make default model region-aware based on AWS_REGION
Move resolveBedrockModelId and helpers from providers.ts to provider-meta.ts
so getDefaultModelId can resolve the correct region prefix (us./eu./ap.) for
the default Bedrock model.

- getDefaultModelId now accepts an optional region parameter
- Falls back to process.env.AWS_REGION when no explicit region is given
- Backend callers pass the project config region where available
- Runtime model resolution in create() still rewrites as before
